telegram_twilio_bot/requirements.txt
python-telegram-bot==20.6
twilio==9.0.1
python-dotenv==1.0.1
